refactor(dispatcher): decouple topology discovery and simplify opcode pattern matching (Phase 3, #979)
